commit 02
deixei mais bonito e fiz as adições que o professor pediu de ultima hora

+++ src/App.java
@@ -1,26 +1,43 @@
+import java.util.List;
+
 public class App {
-   public static void main(String[] args) {
-       iAlunoRepository alunoRepository = new ArrayListAlunoRepository();
+ public static void main(String[] args) {
+   iAlunoRepository alunoRepository = new ArrayListAlunoRepository();
 
-       alunoRepository.adicionar(new Aluno("Lucas", "111"));
-       alunoRepository.adicionar(new Aluno("Maria", "222"));
-       alunoRepository.adicionar(new Aluno("Joao", "333"));
-       alunoRepository.adicionar(new Aluno("Carol", "444"));
-       alunoRepository.adicionar(new Aluno("Eduardo", "555"));
+   alunoRepository.adicionar(new Aluno("Lucas", "1234567890"));
+   alunoRepository.adicionar(new Aluno("Hugo", "2345678901"));
+   alunoRepository.adicionar(new Aluno("Carol", "3456789012"));
+   alunoRepository.adicionar(new Aluno("Manoel", "4567890123"));
+   alunoRepository.adicionar(new Aluno("Victoria", "5678901234"));
 
-       System.out.println("Alunos:");
-       for (Aluno aluno : alunoRepository.listar()) {
-           System.out.println(aluno.getNome());
-       }
-
-       Aluno alunoBuscado = alunoRepository.buscar("Lucas");
-       System.out.println("\nAluno buscado: " + alunoBuscado.getNome());
-
-       alunoRepository.remover(alunoBuscado);
-
-       System.out.println("\nAlunos após remoção:");
-       for (Aluno aluno : alunoRepository.listar()) {
-           System.out.println(aluno.getNome());
-       }
+   System.out.println("Lista de Alunos:");
+   List<Aluno> alunos = alunoRepository.listar();
+   for (Aluno aluno : alunos) {
+       System.out.println(aluno);
    }
+
+   System.out.println();
+
+   Aluno alunoBuscado = alunoRepository.buscar("Lucas");
+   if (alunoBuscado == null) {
+       System.out.println("Aluno não encontrado");
+       return;
+   }
+
+   if (alunoBuscado.getMatricula().length() != 10) {
+       System.out.println("Matricula invalida: " + alunoBuscado);
+       return;
+   }
+
+   System.out.println("Aluno buscado e removido: " + alunoBuscado);
+   alunoRepository.remover(alunoBuscado);
+
+   System.out.println(); 
+
+   System.out.println("Lista de Alunos após remoção:");
+   alunos = alunoRepository.listar();
+   for (Aluno aluno : alunos) {
+       System.out.println(aluno);
+   }
+ }
 }
